Update belongsToPrivateKeys function in util.ts
