+++ test/wallet.spec.ts
@@ -1,113 +1,12 @@
 import * as assert from 'assert';
 import { describe, it } from 'mocha';
-import * as tapyrus from 'tapyrusjs-lib';
 
-import { Balance } from '../src/balance';
-import { Config } from '../src/config';
-import { BaseWallet } from '../src/wallet';
-import { KeyStore } from '../src/key_store';
-import { DataStore } from '../src/data_store';
 import { Utxo } from '../src/utxo';
-import * as util from '../src/util';
+import { createWallet, setUpStub } from './testutil';
 
 import * as sinon from 'sinon';
-class LocalKeyStore implements KeyStore {
-  _wifKeys: string[] = [];
-  _extKeys: string[] = [];
-  network: tapyrus.networks.Network;
-
-  constructor(network: tapyrus.networks.Network) {
-    this.network = network;
-  }
-
-  async addPrivateKey(key: string): Promise<void> {
-    this._wifKeys.push(key);
-  }
-
-  async addExtendedPrivateKey(extendedPrivateKey: string): Promise<void> {
-    this._extKeys.push(extendedPrivateKey);
-  }
-
-  async keys(): Promise<string[]> {
-    return this._wifKeys
-      .map(wif => {
-        return tapyrus.ECPair.fromWIF(wif, this.network).privateKey!.toString(
-          'hex',
-        );
-      })
-      .concat(
-        this._extKeys.map(xpriv => {
-          return tapyrus.bip32
-            .fromBase58(xpriv, this.network)
-            .privateKey!.toString('hex');
-        }),
-      );
-  }
-
-  clear() {
-    this._wifKeys = [];
-    this._extKeys = [];
-  }
-}
-
-class LocalDataStore implements DataStore {
-  utxos: Utxo[] = [];
-
-  async clear(): Promise<void> {
-    this.utxos = [];
-  }
-
-  async add(utxos: Utxo[]): Promise<void> {
-    this.utxos = this.utxos.concat(utxos);
-  }
-
-  async all(): Promise<Utxo[]> {
-    return this.utxos;
-  }
-
-  async balanceFor(keys: string[], colorId?: string): Promise<Balance> {
-    let utxos: Utxo[];
-    const scripts = util.keyToScript(keys, colorId);
-    utxos = this.utxos.filter((utxo: Utxo) => {
-      return scripts.find((script: string) => script == utxo.scriptPubkey);
-    });
-    return util.sumBalance(utxos, colorId);
-  }
-}
-
-const createWallet = (
-  network: string,
-): {
-  wallet: BaseWallet;
-  keyStore: LocalKeyStore;
-  dataStore: LocalDataStore;
-} => {
-  const config = new Config({
-    host: 'example.org',
-    port: '50001',
-    path: '/',
-    network,
-  });
-  const keyStore = new LocalKeyStore(config.network);
-  const dataStore = new LocalDataStore();
-  const wallet = new BaseWallet(keyStore, dataStore, config);
-  return {
-    wallet,
-    keyStore,
-    dataStore,
-  };
-};
 
 describe('Wallet', () => {
-  let stub: sinon.SinonStub;
-  const setUpStub = (wallet: BaseWallet) => {
-    stub = sinon.stub();
-    wallet.rpc.request = (_config: Config, _method: string, _params: any[]) => {
-      const result = stub();
-      return result;
-    };
-  };
-
   afterEach(() => {
     sinon.restore();
   });
@@ -308,7 +207,7 @@ describe('Wallet', () => {
 
     it('save utxos', async () => {
       const { wallet: alice, dataStore } = createWallet('prod');
-      setUpStub(alice);
+      const stub = setUpStub(alice);
       stub.onFirstCall().returns(new Promise(resolve => resolve(unspnets1)));
       stub.onSecondCall().returns(new Promise(resolve => resolve(unspnets2)));
       const wif = 'KzJYKvdPEkuDanYNecre9QHe4ugRjvMvoLeceRr4j5u2j9gEyQ7n';
@@ -371,7 +270,7 @@ describe('Wallet', () => {
     describe('uncolored coin', () => {
       it('get uncolored balance', async () => {
         const { wallet: alice } = createWallet('prod');
-        setUpStub(alice);
+        const stub = setUpStub(alice);
         stub.onFirstCall().returns(new Promise(resolve => resolve(unspnets)));
         await alice.importWif(wif);
         await alice.update();
@@ -384,7 +283,7 @@ describe('Wallet', () => {
     describe('colored coin', () => {
       it('get colored balance', async () => {
         const { wallet: alice } = createWallet('prod');
-        setUpStub(alice);
+        const stub = setUpStub(alice);
         stub.onFirstCall().returns(new Promise(resolve => resolve(unspnets)));
         await alice.importWif(wif);
         await alice.update();
